+++ lib/rage/cable/cable.rb
@@ -8,11 +8,11 @@ module Rage::Cable
   #     run Rage.cable.application
   #   end
   def self.application
-    protocol = Rage.config.cable.protocol
-    protocol.init(__router)
+    # explicitly initialize the adapter
+    __adapter
 
-    handler = __build_handler(protocol)
-    accept_response = [0, protocol.protocol_definition, []]
+    handler = __build_handler(__protocol)
+    accept_response = [0, __protocol.protocol_definition, []]
 
     application = ->(env) do
       if env["rack.upgrade?"] == :websocket
@@ -32,6 +32,15 @@ module Rage::Cable
   end
 
   # @private
+  def self.__protocol
+    @__protocol ||= Rage.config.cable.protocol.tap { |protocol| protocol.init(__router) }
+  end
+
+  def self.__adapter
+    @__adapter ||= Rage.config.cable.adapter
+  end
+
+  # @private
   def self.__build_handler(protocol)
     klass = Class.new do
       def initialize(protocol)
@@ -42,33 +51,22 @@ module Rage::Cable
         end
 
         @protocol = protocol
+        @default_log_context = {}.freeze
       end
 
       def on_open(connection)
-        Fiber.schedule do
-          @protocol.on_open(connection)
-        rescue => e
-          log_error(e)
-        end
+        connection.env["rage.request_id"] ||= Iodine::Rack::Utils.gen_request_tag
+        schedule_fiber(connection) { @protocol.on_open(connection) }
       end
 
       def on_message(connection, data)
-        Fiber.schedule do
-          @protocol.on_message(connection, data)
-        rescue => e
-          log_error(e)
-        end
+        schedule_fiber(connection) { @protocol.on_message(connection, data) }
       end
 
       if protocol.respond_to?(:on_close)
         def on_close(connection)
           return unless ::Iodine.running?
-
-          Fiber.schedule do
-            @protocol.on_close(connection)
-          rescue => e
-            log_error(e)
-          end
+          schedule_fiber(connection) { @protocol.on_close(connection) }
         end
       end
 
@@ -82,6 +80,15 @@ module Rage::Cable
 
       private
 
+      def schedule_fiber(connection)
+        Fiber.schedule do
+          Thread.current[:rage_logger] = { tags: [connection.env["rage.request_id"]], context: @default_log_context }
+          yield
+        rescue => e
+          log_error(e)
+        end
+      end
+
       def log_error(e)
         Rage.logger.error("Unhandled exception has occured - #{e.class} (#{e.message}):\n#{e.backtrace.join("\n")}")
       end
@@ -94,10 +101,14 @@ module Rage::Cable
   #
   # @param stream [String] the name of the stream
   # @param data [Object] the object to send to the clients. This will later be encoded according to the protocol used.
+  # @return [true]
   # @example
   #   Rage.cable.broadcast("chat", { message: "A new member has joined!" })
   def self.broadcast(stream, data)
-    Rage.config.cable.protocol.broadcast(stream, data)
+    __protocol.broadcast(stream, data)
+    __adapter&.publish(stream, data)
+
+    true
   end
 
   # @!parse [ruby]
@@ -119,6 +130,11 @@ module Rage::Cable
   #     def close
   #     end
   #   end
+
+  module Adapters
+    autoload :Base, "rage/cable/adapters/base"
+    autoload :Redis, "rage/cable/adapters/redis"
+  end
 
   module Protocol
   end