+++ lib/rage/middleware/reloader.rb
@@ -2,15 +2,32 @@
 
 class Rage::Reloader
   def initialize(app)
+    Iodine.on_state(:on_start) do
+      Rage.code_loader.check_updated!
+    end
     @app = app
   end
 
   def call(env)
-    Rage.code_loader.reload
-    @app.call(env)
+    with_reload do
+      @app.call(env)
+    end
   rescue Exception => e
     exception_str = "#{e.class} (#{e.message}):\n#{e.backtrace.join("\n")}"
     puts(exception_str)
     [500, {}, [exception_str]]
+  end
+
+  private
+
+  def with_reload
+    if Rage.code_loader.check_updated!
+      Fiber.new(blocking: true) {
+        Rage.code_loader.reload
+        yield
+      }.resume
+    else
+      yield
+    end
   end
 end