+++ spec/openapi/builder/summary_spec.rb
@@ -161,5 +161,134 @@ RSpec.describe Rage::OpenAPI::Builder do
         subject
       end
     end
+
+    context "with empty commments" do
+      let_class("UsersController", parent: RageController::API) do
+        <<~'RUBY'
+          # Returns the list of all users.
+          #
+          # 
+          def index
+          end
+        RUBY
+      end
+
+      let(:routes) do
+        { "GET /users" => "UsersController#index" }
+      end
+
+      it "returns correct schema" do
+        expect(subject).to eq({ "openapi" => "3.0.0", "info" => { "version" => "1.0.0", "title" => "Rage" }, "components" => {}, "tags" => [{ "name" => "Users" }], "paths" => { "/users" => { "get" => { "summary" => "Returns the list of all users.", "description" => "", "deprecated" => false, "security" => [], "tags" => ["Users"], "responses" => { "200" => { "description" => "" } } } } } })
+      end
+
+      it "does not log error" do
+        expect(Rage::OpenAPI).not_to receive(:__log_warn)
+        subject
+      end
+    end
+
+    context "after another tags" do
+      let_class("UsersController", parent: RageController::API) do
+        <<~'RUBY'
+          # @deprecated
+          # Returns the list of all users.
+          #
+          # @internal this is an internal comment
+          def index
+          end
+        RUBY
+      end
+
+      let(:routes) do
+        { "GET /users" => "UsersController#index" }
+      end
+
+      it "returns correct schema" do
+        expect(subject).to eq({ "openapi" => "3.0.0", "info" => { "version" => "1.0.0", "title" => "Rage" }, "components" => {}, "tags" => [{ "name" => "Users" }], "paths" => { "/users" => { "get" => { "summary" => "Returns the list of all users.", "description" => "", "deprecated" => true, "security" => [], "tags" => ["Users"], "responses" => { "200" => { "description" => "" } } } } } })
+      end
+
+      it "does not log error" do
+        expect(Rage::OpenAPI).not_to receive(:__log_warn)
+        subject
+      end
+    end
+
+    context "as the last tag" do
+      let_class("UsersController", parent: RageController::API) do
+        <<~'RUBY'
+          # @deprecated
+          # @internal this is an internal comment
+          #
+          # Returns the list of all users.
+          def index
+          end
+        RUBY
+      end
+
+      let(:routes) do
+        { "GET /users" => "UsersController#index" }
+      end
+
+      it "returns correct schema" do
+        expect(subject).to eq({ "openapi" => "3.0.0", "info" => { "version" => "1.0.0", "title" => "Rage" }, "components" => {}, "tags" => [{ "name" => "Users" }], "paths" => { "/users" => { "get" => { "summary" => "Returns the list of all users.", "description" => "", "deprecated" => true, "security" => [], "tags" => ["Users"], "responses" => { "200" => { "description" => "" } } } } } })
+      end
+
+      it "does not log error" do
+        expect(Rage::OpenAPI).not_to receive(:__log_warn)
+        subject
+      end
+    end
+
+    context "after a text tag" do
+      let_class("UsersController", parent: RageController::API) do
+        <<~'RUBY'
+          # @description this is a test description
+          # Returns the list of all users.
+          def index
+          end
+        RUBY
+      end
+
+      let(:routes) do
+        { "GET /users" => "UsersController#index" }
+      end
+
+      it "returns correct schema" do
+        expect(subject).to eq({ "openapi" => "3.0.0", "info" => { "version" => "1.0.0", "title" => "Rage" }, "components" => {}, "tags" => [{ "name" => "Users" }], "paths" => { "/users" => { "get" => { "summary" => "Returns the list of all users.", "description" => "this is a test description", "deprecated" => false, "security" => [], "tags" => ["Users"], "responses" => { "200" => { "description" => "" } } } } } })
+      end
+
+      it "does not log error" do
+        expect(Rage::OpenAPI).not_to receive(:__log_warn)
+        subject
+      end
+    end
+
+    context "after a multi-line tag" do
+      let_class("UsersController", parent: RageController::API) do
+        <<~'RUBY'
+          # @description this
+          #   is
+          #   a
+          #   test
+          #   description
+          # Returns the list of all users.
+          def index
+          end
+        RUBY
+      end
+
+      let(:routes) do
+        { "GET /users" => "UsersController#index" }
+      end
+
+      it "returns correct schema" do
+        expect(subject).to eq({ "openapi" => "3.0.0", "info" => { "version" => "1.0.0", "title" => "Rage" }, "components" => {}, "tags" => [{ "name" => "Users" }], "paths" => { "/users" => { "get" => { "summary" => "Returns the list of all users.", "description" => "this is a test description", "deprecated" => false, "security" => [], "tags" => ["Users"], "responses" => { "200" => { "description" => "" } } } } } })
+      end
+
+      it "does not log error" do
+        expect(Rage::OpenAPI).not_to receive(:__log_warn)
+        subject
+      end
+    end
   end
 end