+++ spec/rage/request_spec.rb
@@ -4,6 +4,7 @@ RSpec.describe Rage::Request do
       "rack.upgrade?" => nil,
       "rack.upgrade" => nil,
       "rack.version" => [1, 3],
+      "rage.request_id" => "5zfcwiy09bary01l",
       "SCRIPT_NAME" => "",
       "rack.url_scheme" => "http",
       "HTTP_VERSION" => "HTTP/1.1",
@@ -64,7 +65,7 @@ RSpec.describe Rage::Request do
     env.delete("HTTP_USER_AGENT")
     expect(request.user_agent).to be_nil
   end
-
+  
   it "handles the host property of a request" do
     expect(request.host).not_to be_nil
   end
@@ -119,5 +120,9 @@ RSpec.describe Rage::Request do
 
   it "handles the format property of a request" do
     expect(request.format).not_to be_nil
+  end
+  
+  it "returns the correct request ID" do
+    expect(request.request_id).to eq("5zfcwiy09bary01l")
   end
 end